requirements.txt
# Enhanced Football Tracking System Dependencies
# BoT-SORT + SigLIP/UMAP Team Classification

# Core Vision
opencv-python>=4.8.0
numpy>=1.24.0

# YOLO Detection & BoT-SORT Tracking
ultralytics>=8.0.0

# SigLIP Team Classification
transformers>=4.35.0
torch>=2.0.0
torchvision>=0.15.0

# UMAP Dimensionality Reduction
umap-learn>=0.5.4

# Clustering & ML
scikit-learn>=1.3.0

# Data Processing & Export
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Network Analysis
networkx>=3.1

# Configuration
pyyaml>=6.0

# Optional: Accelerators
# sentence-transformers>=2.2.0  # For alternative embeddings
# faiss-gpu>=1.7.2  # For fast similarity search